pyproject.toml
[project]
name = "runeform"
version = "0.3.0"
description = "Compositional layout engine for brand marketing - POC 3"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "pydantic>=2.0",
    "pillow>=11.0",
    "anthropic>=0.52",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.2.2",
    "z3-solver>=4.12",
    "cairosvg>=2.7",
    "spandrel>=0.4",
    "torch>=2.0",
    "skia-python>=120",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "playwright>=1.40",
    "pytest-playwright>=0.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
